feat: switch Url for Uri in Destination with custom Option serde impl
The spec allows for sink parameters to specify both a ref and a uri. In
that case, the uri must be a path that is set on the resolved ref url.
Only if no ref is specified must the uri be an absolute url.

This makes the url type too strict for this field since url must always
be absolute.